refactor: Hide `base::pretty_print` module.
* Move `base::pretty_print` to `base::types::pretty_print`.
* Re-exports `base::types::Printer`.
* `gluon_format` crate uses `base::types::Printer` in a newly created
newtype struct.
